requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
motor==3.1.2
pydantic==1.10.7
passlib==1.7.4
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.0
email-validator==2.0.0
python-multipart==0.0.6